<feat>: 단체 생성 & 조회 API 구현

+++ src/main/java/org/unisphere/unisphere/group/dto/response/GroupMemberListResponseDto.java
@@ -20,5 +20,5 @@ public class GroupMemberListResponseDto {
 	private final List<GroupMemberDto> groupMembers;
 
 	@Schema(description = "전체 회원 수", example = "100")
-	private final int totalMemberCount;
+	private final long totalMemberCount;
 }
